ENGR00235687 No alarm from HDMI in JB
JB will output alarm to HDMI and speaker in the same time.
This is caused by HDMI's period_size is fixed in the driver, the period_size
set int HAL will be changed in pcm_open. Then the pre-set threshold
is not suitable for HDMI, which will cause the HDMI output mute. remove the
threshold, then tinyalsa will set it according the actual period_size.

Change-Id: I33ce392f4113f7e59ae7fbdab1faec4db1683d85
Signed-off-by: b02247 <b02247@freescale.com>