tinyalsa: allow HDMI to be used as primary output/input device
This patch allows the HDMI audio driver (imx-hdmi-soc) to be
used as the primary audio output when the persistent property
persist.audio.hdmi_out is set to 1.

It also allows to use the  the TC358743 HDMI to MIPI chip to be
used as primary audio input devices when the property
persist.audio.hdmi_in is set to 1.

Note that this is not completely dynamic, because the implementation
checks for the property when opening the audio device. In other words,
you can't switch in the middle of audio playback.

You can switch the value of the property from 0 to 1 at run-time,
but the expected use-case is that the audio routing will be set
at boot time based on what's connected to the board.

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>
Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>