MA-9135 Fix CTS android.view.cts.SurfaceViewSyncTests failed.
When run android.view.cts.SurfaceViewSyncTests CTS, GPU hang with AXI bus error.
The root cause is virual display target buffer is not a real physical address.
In this case GPU MMU is enabled and physcial address is a GPU mapped
virtual address which is a different value in different process.
Call hwc_lockSurface to get physical address in this process.

Change-Id: Ic82b3151f3741fa5033a960cb8b15bb44f349589
Signed-off-by: ivan.liu <xiaowen.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1397
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>