MA-9695 Fix many error mesages issues when plug in mouse.
When plug in mouse, there are many error messages like "Failed to set cursor
position to (153, 411): Unsupported".This is because layer try to all
SetCursorPosition when detect layer CompositionType is Cursor, but failed.
We don't support SetCursorPosition. And if we don't support SetCursorPosition,
we need modify CompositionType from Cursor to Device or Client.

Change-Id: I7219dca3b20f58bef1b5083fc430042043a1a9b9
Reviewed-on: http://androidsource.ap.freescale.net/project/2064
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>