MA-10343 USB camera crash when take picture on Android.O.
Sometimes mChange can't be guaranteed is true when run handlConfigureLocked().
When mChange is false, old buffers can't be freed. Camers hal won't allocate new
buffers. If the number of buffers is not match with request, process will crash.
Move assignment operation for mChange from configure() to handlConfigureLocked()
to make sure mChange is true when need to do new configure operation.

Change-Id: I02e1f63cca534ba9dcfba6cb9673fcfe0820c33a
Signed-off-by: Wenbo.Zhu <wenbo.zhu_1@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/2943
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>