MA-10811 [Android_OR] Preview blocked after take several pictures
Occurs on device has no IPU/PXP(mscale), err log as:
"Software don't support resize"

The v4l2 use differet resolution as preview stream required.
It's due to the preview captrue request dropped unexpectly.

To ehance performance, there's a resolution/fps/format judge in
the command thead, if same, drop it.

But the v4l2 resolution/fps/format is changed in the work thread.
So there's a un-consistency. Should move the judgement to the work thread.

Change-Id: Ie46882f77b1edbb55daeafa44b7445d455df8d3b
Signed-off-by: fang hui <hui.fang@nxp.com>