MA-12566 Enable usb rndis function using usb gadget hidl
Rndis(tethering)interface, usb0, registration/deregistration
is broken.If a user try to switch to other functions or
disable usb tethering or unplug the usb cable, it doesn't
kill "usb0" interface.

Clean registration and deregistration can be done by creating
the rndis function when user enable tethering from Settings'usb
preference or usb tethering, and deleting the function when we
switch to other function or disable tethering or unplug the cable.

Change-Id: I3e212ef3ba67942c3ccb90cc4b61c0611892ea9c
Signed-off-by: yang.tian <yang.tian@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/5777
Reviewed-by: Wang Haoran <elven.wang@nxp.com>