MA-12881 ADB:adb offline during STS/VTS test
When adbd init successfully, "sys.usb.ffs.ready" property is set 1.

So when doing usb function set operation, firstly the property need to
be cleaned, or it will trigger bind f_fs without adbd ready.
And then we also need to waiting for "sys.usb.ffs.ready" property to be
1 after trigger adbd start which indicates adbd is actually ready.

Change-Id: Iaa2af3cd500747278c60e773157943d093a72a62
Signed-off-by: yang.tian <yang.tian@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/6193
Reviewed-by: Wang Haoran <elven.wang@nxp.com>