MA-13920-3 [Android] enable vehicle hal based on the info from mcu
create socket with type AF_NETLINK which communicate with kernel.
the server addr need to filled with pid when bind socket.
it use sendto/recvfrom to write/read message to kernel rpmsg driver.
the default code use tcp to communicate with simulator.
the socket can't been created successfully in class hal. So start it
when boot completed.

Change-Id: I7572caf4b167afe0e241e0eeedfa4d95a8811c59
Signed-off-by: zhang sanshan <sanshan.zhang@nxp.com>