MA-14176 [Android Pi] On LPA play, lots of start_output_stream_primary log.
On the begining of LPA, the button click sound also sent to HAL via primary
output. But the device is already occupied by LPA direct output. The log
continues until the click sound "consumed".
To fix it, show the log only when the primary output has chance to open
the sound card.

Change-Id: Iacb8ea520fb2d12a7d389cace81d98d1f95d2012
Signed-off-by: fang hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7529
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>