MA-14900 fix "Broken pipe" issue in VTS
Some modules in VTS will fail with "Broken pipe" prompt for its cases.
This happens when when the module preparation finishes and start to test
cases by order, then adbd restarted, cause TCP socket forward path
lost, following transactions will cause "Broken pipe"

This patch changes the logic in "setCurrentUsbFunctions" function, usb
configfs is only torn down and brought back up when  the "functions"
to be set is different from the "mCurrentUsbFunctions", or last operation
of "setCurrentUsbFunctions" fails to set the gadget function.

In fact, the cases fail with the prompt of "'int' object is not iterable"
has the same reason, so this workaround eleminate this kind of errors.

USB hal doesn't have the permission to change sys.usb.ffs.ready. remove
the line which try to change it.

Change-Id: I5588d622dd7e1da3c0f8658cb1eef1f9e39eac59
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/8587
Reviewed-by: Tian Yang <yang.tian@nxp.com>