Fix jpeg transport header offset
As GPU can't recognize BLOB, the format is faked to YUYV. The size of
the allocated graphic buffer is slightly larger than jpeg buffer size
because of memory alignment. This results in jpeg transport header being
written to a position different from the one expected in
Image_getJpegSize() in android_media_Utils.cpp, causing the camera
picture size to be always set to the maximum value.

Signed-off-by: Laura Nao <laura.nao@kynetics.com>
Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>