MA-13920-3 [Android] enable vehicle hal based on the info from mcu
create socket with type AF_NETLINK which communicate with kernel.
the server addr need to filled with pid when bind socket.
it use sendto/recvfrom to write/read message to kernel rpmsg driver.
the default code use tcp to communicate with simulator.
the socket can't been created successfully in class hal. So start it
when boot completed.

Change-Id: I8c075e6e84c93ab99c1b4cf2909ae11dd35678cb
Signed-off-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/10008
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>