MA-15575-1 Add support for oemlock 1.0 hal
Add oemlock 1.0 hal support, this implementation help enhance the system
security because it will store the oem device unlock states to secure
storage(RPMB) instead of normal storage (e.g. persistdata part).

The carrier unlock is used to implement features like "sim lock" or
"phone lock" so the carrier can determine if the device is allowed to
be unlocked. It will need process to authenticate the device and depend
on some info like the Medel ID. On i.MX platforms, we don't support this
feature for now and the implementation will always reture ok.

Test: Oem device unlock turns on/off.

Change-Id: I87aa9441bc44dde0ee172f7f52676efd292f906d
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/10206
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Wang Haoran <elven.wang@nxp.com>