MA-16941[Android_AUTO] Some displays cannot show normally on mek-8qm with car image
Some displays are not ready or the connection is false. So these displays are
not reported to framework when register Hotplug callback function in
hwc2_register_callback.
But these displays will be ready when PollFileThread enumerate display again and
the connection has change to true. The handleKmsHotplug handler will not report
to framework if display connection is true and not change.

Change-Id: If77f801aa535931d4da8f3ec0777555c873e9a88
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/11905
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>