MA-17033 [Android_8QM/8QXP_MEK]Camera: In Files APK thumbnails of camera captured pictures are green.(100%)
The thumbnails in right-up side is ok when take picture. The picture shown in Gallery is also ok.
However the cover of picture in File apk is green, the picture is also ok after open the picture in file.

It will do resize if the jpeg input size is not match output size, however yuvResize in HwJpegEncoder.cpp is null by default.
implement yuvResize by default in YuvToJpegEncoder. It is inherited by Yuv422IToJpegEncoder/HwJpegEncoder.

Change-Id: Ib09a1cc1e2328c360ca8f948c42f6da9bfba9922
Signed-off-by: zhang sanshan <pete.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12087
Reviewed-by: fang hui <hui.fang@nxp.com>
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>