MA-17238 Android 10 CTS AudioRecordTest#testTimestamp failed
Failed as "expect maxAbsJitterMs(10.441515006821282) < 6.0".

The test use 23456 sample rate to record 2 seconds. Every time read 512 frames,
It will get AudioTimestamp ts, which has member framePosition and nanoTime.
Then it subs the lastTs to get deltaTimeNs, deltaFrames.
Divide deltaFrames by sameple rate to get deltaFramesNs.
jitterMs is the jitter between deltaTimeNs, deltaFramesNs.

In HAL, the record sample rate is 48000, period size is 512.
The frames captured is update by the unit of period size. when DMA finished,
frames wll eacrease 512. So if not get timestamp right after DMA finish,
in theory, the frames may has deviation max to 512, which leads to the max
time (frames/rate) deviation as 512 * 1000 / 48000 = 10.67 ms.

The test and HAL sampe rate (23456 vs 48000) are not integer multiples to
each other, which extreamly exposes the deviation risk.

To fix it, half the period size to 256 and double the period count to 32.

Change-Id: Ieb0b35dc94703d105be4ef5521d19d53ef4e2c51
Signed-off-by: fang hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12561
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>