MA-17419[Android] Use a new property to turn on the second display HDMI-A-2
The second display is forced off by bootargs "video=HDMI-A-2:d" to save
some boot time in car image. Then it will be turn on in init.rc when boot completed.
But in android R, the display driver is not ready when boot_completed.
The PollFileThread thread monitor the display device status. When display
device is ready, it will enumerate these display devices. Here set the property
to trigger the init.rc to turn on the second display.

Change-Id: I728ead5821c9586416e6b05250fcb93534d4ad31
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12636
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>
(cherry picked from commit 95dfb5a072d16c33368671259784ee0f0f16045c)
Reviewed-on: http://androidsource.ap.freescale.net/project/12652
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>