MA-17141[Android_8QXP/8QM_MEK_AUTO]Display: Camera display will flash for a second after device boot up.(100%)
As the patch "MA-16771 Fix window animation issue with DPU 2D HWComposer" force
to use GPU3D to do compose when the global alpha is not 0xff for
pixel alpha + blending + global alpha case, while the EVS need to use
HWC to do composition of camera frame and UI layer, it cause screen switch to UI
and switch back to EVS frame sometimes when EVS start. When system boot up, the
alpha of the UI is not always 0xff. It cause UI showing when EVS is working.

Here force to use HWComposer when there are HwLayers.

Change-Id: I2ae7b904cff555607361d0d01a28b248b11fae2d
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12737
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>