Revert "MA-12881 ADB:adb offline during STS/VTS test"
This reverts commit a2fd5405de611d0451f109375e8f4aa5484805e0.

The reverte commit tries to restart adbd when switch usb functions, it
is a workaround of usb offline issue happens on evk_7ulp in Android
OR8.1.

Now in Android11, adbd is not restarted because of selinux rules, and
there would be much denial message output on console with this patch.

Change-Id: I47167ddfa74a70e7af229ce123ae81bafe39b3a4
Reviewed-on: http://androidsource.ap.freescale.net/project/13239
Reviewed-by: Tian Yang <yang.tian@nxp.com>