MA-17878[Android] Refine display HAL to make system boot up without depending on primary_display and fbTileSupport
As the primary_display and fbTileSupport bootargs parameters need to change
when use lcdif diver on evk_8mq, here refine the code to avoid depend on
such bootargs.
If the primary_display is not match the drm version, will use the first
available conncected display as primary display.
If the display driver don't support frame buffer tile format, will set the
hwc.drm.fbTileSupport property to indicate not use tile format in framework
and GPU driver.

Change-Id: I7986a15081b4ac4824baebc7d2c8b23e626d0886
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13519
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>