MA-17938 [Android 11] GCH: black frames around 5M picture
Previous code judeg capture intent in ConfigurePipeline(), but no intent
is passed in the function(). Then V4l2 is confiured by preview stream(1024x768),
then resize to 2592x1944 to take picture, so there are black frames.

To fix it, judeg capture intent in SubmitRequests().
Also, to make it easy to use, add ConfigAndStart() and Stop() in VideoStream.

ConfigurePipeline/SubmitRequests are somewhat like
configureStreams/processCaptureRequest in libcamera3.

Signed-off-by: fang hui <hui.fang@nxp.com>

Change-Id: I059097dcce5c6c89512451cf1d7c0c9a1bfc30c5
Reviewed-on: http://androidsource.ap.freescale.net/project/13635
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>