MA-18165 [Android 11] testRepeatingJpeg failed
Failed as "junit.framework.AssertionFailedError: expected:<640> but was:<-1>".

The test only pass JPEG related meta in the first request for 320x240 picture.
In previous code, pass each capture request's meta to HandleFrameLocked().
So when 640x480, JPEG related meta is NULL, so the HAL will not encode JPEG.

To fix it, need record the latest valid meta from requests, and pass it to
HandleFrameLocked().

Change-Id: I700a13ee5ea5a1d0e037cdf6aa5a3add009cd4ba
Signed-off-by: fang hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13887
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>