MA-18408[Android] EVS cannot work after enable gralloc v4
The EVS related processes use the MemoryManager class directly and the allocated
buffer handle will be shared through binder and native_handle will be check.
In gralloc v4, the numFds and numInts are set in gralloc_driver class which is a
wapper of MemoryManager.
So need to set the default value to numFds and nunInts for gralloc v4 in case the
gralloc_driver APIs are not used.
error log:
[   60.471188] binder_alloc: 2526: binder_alloc_buf size 10387409968 failed, no address space
[   60.479502] binder_alloc: allocated: 2648 (num: 1 largest: 2648), free: 1037736 (num: 1 largest: 1037736)

Change-Id: Ic27b851288a7340edd9457a1d68dba966bc3a9ae
Changr-Id: Ib9f5aeeb83f1988f898bf3f61b95df30c97f1b76
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14275
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>