MA-18523-2 remove light hal
in LightsService.java, we have 2 ways to set Brightness, one is via display hal,
another one is via light hal,

there are 3 judging conditions, the first condition is "brightnessMode == BRIGHTNESS_MODE_USER",
from the code, we don't need to support VR, so we will always use BRIGHTNESS_MODE_USER ,
the second condition is "!shouldBeInLowPersistenceMode()", this is about nLowPersistence in VR mode.
the third condition is "mSurfaceControlMaximumBrightness == 255", this is to judge Display
brightness support, we actually implement the Display::initBrightness,
so we can meet the 3 judging conditions.

so in our default image, use display hal to setDisplayBrightness instead of using light hal
to set Brightness, so even add aidl light hal, will still use the SurfaceControl to
setDisplayBrightness, and I confirm that when we use light sensor to adjust backlight,
also follow the SurfaceControl way.

and go through the light service and implements, all implemented in display hal,
so light hal is not needed now.

Change-Id: I163706a4b1349d40f23b71e3744f25ef75577e4e
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14557
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>