MA-18800 Fix coverity issues of camera
Fix camera hal coverity issues as below:

11903832 Uninitialized scalar field
Non-static class member is not initialized in this
constructor nor in any functions that it calls.

11903765 Empty virtual destructor
android::MMAPStream has a destructor and a pointer to it is upcast to
class android::ImxStream which doesn't have a virtual destructor.

11903852 Unchecked return value
Calling parse without checking return value.

11903744 Unchecked return value
Calling ioctl without checking return value.

11903887 Resource leak
Variable vidDir going out of scope leaks the storage it points to.

11903881 Uninitialized pointer field
Non-static class member is not initialized in this constructor nor
in any functions that it calls.

11903876 Uninitialized scalar field
Non-static class member is not initialized in this constructor nor
in any functions that it calls.

12010912 Resource leak
Overwriting node in node = nodes leaks the storage that node points to.

11903881 Uninitialized scalar field
Non-static class member is not initialized in this constructor nor
in any functions that it calls.

11903875 Uninitialized scalar field
Non-static class member is not initialized in this constructor nor
in any functions that it calls.

11903850 Resource leak
Variable pipeline_info going out of scope leaks the storage it points
to.

11903812 Resource leak
Variable pipeline_request going out of scope leaks the storage it points
to.

11903771 Uninitialized pointer field
Non-static class member is not initialized in this
constructor nor in any functions that it calls.

Change-Id: Iac3892ed94730a535f7e0760407c2c8dd91275bb
Signed-off-by: yang.tian <yang.tian@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15354
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>