MA-18654-2 eleminate the log of UDC not found of busy
When the usb port dose not work as a device, UDC device does not exist.
If try to bind the UDC with configfs, there will be log as below:

  udc-core: couldn't find an available UDC or it's busy

inotify watch on "/sys/class/udc/" does not work, inotify watch on the
sysfs path of usb controller can work, that path is different with
different boards/SoCs, and may change from one kernel version to
another.

This commit checks whether the link under "/sys/class/udc" can be
accessed, if it can, then bind the udc with configfs.

Change-Id: Ice3dd2c8323d6eadde2ddd23cae97d725d198129
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15395
Reviewed-by: Wang Haoran <elven.wang@nxp.com>