MA-19069 bring up Android12 on imx devices
This patch tries to make the build process of target evk_8mm-eng can be
executed to the stage that .ninja file is generated and ninja can be
started to build source code. Source code compilation errors are not
fixed in this commit.

"PlatformSdkVersionInt()" is changed to "PlatformSdkVersion().FinalOrFutureInt()"
in the repo under build/soong, make similar changes in:
    alsa/alsa.go
    camera/camera.go
    evs/evs_hal/evs.go

the module "android.hardware.power-ndk_platform" now does not exist, refer to
the changes in hardware/google/pixel/thermal/Android.bp, change it to be
"android.hardware.power-V1-ndk_platform" in:
    power/Android.bp

the module "libqemu_pip" is not defined in android12 preview code, its
source code is moved from "system/core/qemu_pipe/" to
"hardware/interfaces/automotive/vehicle/2.0/default/impl/vhal_v2_0".
In this commit, automotive is not supported temporarily, directly remove
"libqemu_pip" in:
    vehicle/Android.bp

code in "system/core/base" is moved to "system/libbase/include", so
make corresponding path changes in:
    alsa/Android.bp

Change-Id: If30bf13e1bb318cf46fd84d75ca41e9c737449cc
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/16632
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>