MA-19575-1 audio: Fix coverity issue of using invalid iterator
This patch fix coverity issue CID: 17918077
Using invalid iterator: Dereferencing iterator "config.end()"
though it is already past the end of its container.

&*config.end() returns an iterator that points to the next position
of the last element of the container. when it is already past the end
of its container &*config.end()-1, &*config.end() should not be
Dereferenced again.

use string data() function to get array from string config, and parse
each element in the array.

Change-Id: If6e01440df5d8c923f8899eaaeb29474d1f71d1f
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17322
Reviewed-by: dandan.sun <dandan.sun@nxp.com>