MA-19666 android.hardware.camera2.cts.NativeStillCaptureTest#testStillCapture failed on 8mn on android-12
The test take 10 pictures in 5 seconds.
The test use max size picture on android-12, while 480p on android-11.
So it failed on 8mm/8mn/8mq/8ulp.

In test code, if reported "minFrmDuration x 10 >  5 seconds", will use the
reported minFrmDuration to set the timeout. So we add "minframeduration_blob_5M"
in HAL for related boards.

Change-Id: I9fed2459c0594206a3a72ed1bc6666d9458174f7
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17497
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>