MA-19727[Android_I.MX8MQ_EVK]Video:The playback screen is black when play the youtube stream with chrome.(90%)
In some cases, the video frame is in a single layer and HWC regrad it as
overlay layer, will use overlay plane to show it. Sometimes some applicatons
like Chrome has "Background for SurfaceView[xxx]" layer and composite into UI.
This layer is not transparent will cover the video frame. So the video area
show black.
Make such layer as DEVICE type and so it will not composite to UI to avoid
such issue.

Change-Id: I071acf72f35773bf2586ec502dedfeea2227d1a5
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17691
Reviewed-by: juan.hao <juan.hao@nxp.com>