MA-17872-1 check udisk mount path based on each enty in the path
This storage_test now checks whether the udisk mount path in fstab is
valid. Previously, the mount path is in the form as below, this test
module erases the last "*" symbol, and checks whether the remaining path
can be accessed.

    /devices/platform/soc@0/32c00000.bus/32e40000.usb/*

On evk_8ulp, the mount path is now as below, regex is used in two
entries in the path. To make this test module works, it is changed to
check the path from the begin, access the directory, and read entries in
that directory, and match with the entry in this path. If this mount
path can be matched with a real path, then the test succeeds.

    /devices/platform/soc@0/29800000.bus/299[02]0000.usb/*

This test has below assumptions:
  1. the mount path starts with "/"
  2. regex is used inner an entry not inter entries.

Change-Id: I95462b037a9b114529594637ab578a2a8f933dea
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17721
Reviewed-by: Zhai.He <zhai.he@nxp.com>