MA-20181 Dynamically switch the plane of Android UI.
The Android UI and video need to be displayed on the overlay plane
since secure ui need to be displayed on the primary plane so the
indexes submitted to the DCSS need to be swapped.

And when secure ui display, both video and Android UI should be
composerd by using the GPU, so checkOverlay should return false.

And use the "ro.boot.androidui.overlay" property to judge whether
use overlay plane to display androidui or use primary plane.
When use overlay to display android ui, when from secure ui back to
android ui, we need to use force_modeset to flush dcss reg force.

Change-Id: I6923fe0b163b4e08a47a64508b8345c3bf3b7230
Signed-off-by: zhai.he <zhai.he@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/18349
Reviewed-by: zhang bo <bo.zhang@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>