MA-20311 find boot device name with another property
The property ro.boot.boot_device_root is removed in android12.0.0_2.0.0
rc1 release. This property was based on a boot config passed by the
uboot to kernel.

In this patch, we rely on below properties which are based on the files
under the directory "/proc/device-tree/firmware/android", they correspond
to nodes in the device tree. Take evk_8mn as an example, the two nodes in
the device tree are:

    boot_devices_mmcblk1 = "soc@0/30800000.bus/30b50000.mmc";
    boot_devices_mmcblk2 = "soc@0/30800000.bus/30b60000.mmc";

When boot up the android, the two nodes is converted to two properties as
below:

    [ro.boot.boot_devices_mmcblk1]: [soc@0/30800000.bus/30b50000.mmc]
    [ro.boot.boot_devices_mmcblk2]: [soc@0/30800000.bus/30b60000.mmc]

Check the value of these properties and compare to the uevent sub path
of the boot device, then the boot device name can be retrieved.

Change-Id: Idde6a4207167ba3456fd1aceef10622135ae3b99
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19066
Reviewed-by: ji.luo <ji.luo@nxp.com>