MA-19770 Won't use overlay when the video is on the top of the UI.
After opening Toca Kitchen apk in 8mq, the screen display was white.
Root cause is for the apk, video is on the top of UI.
For 8mq, primary plane(UI) is on the top of the overlay plane(video).
Can not handle such case in overlay, video will not be displayed correctly.
Here check the layer zorder, and fall back to 3D if detect video is on the
top of UI.

Change-Id: I72ce544ea1fde10cb609fe6ceb4a0a01afa2c5f0
Signed-off-by: Jessie Hao <juan.hao@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19316
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-by: zhang bo <bo.zhang@nxp.com>