MA-19770-2 Fix secure video black issue in split screen mode.
The secure video is black in split screen mode.
This can happen when the other screen is normal UI or normal video playback
or when clicking secure video UI.
The previous patch checks the layer z-order and falls back to 3D if detect
video is on the top of the UI.
Should consider the displayFrame along with z-order.
If the video is on the top of the UI, but UI displayFrame does not contain
the video, can still use overlay.

Change-Id: Ia335de3d3397dc75deb2b7662b5624b96640128a
Signed-off-by: Jessie Hao <juan.hao@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19889
Reviewed-by: Zhai.He <zhai.he@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
(cherry picked from commit ad1e65691a462234c0c4d8ec7fe77cc6e1a8aa5a)
Reviewed-on: http://androidsource.ap.freescale.net/project/19892
Reviewed-by: zhang bo <bo.zhang@nxp.com>
Reviewed-by: Hui Wang <hui.wang_7@nxp.com>