MA-20728 android.hardware.camera2.cts.BurstCaptureTest#testJpegBurst failed on os08a20
The test expect that sent a request, then can get it's result in the next "maxPipelineDepth" results.
Besides pipeline stages, the camera HAL can also hold requests in a list.
After several tests, with os08a20, the 4th or 5th result is the expected one.
To fix the issue, set ANDROID_REQUEST_PIPELINE_MAX_DEPTH to 8.

Change-Id: Ib36e2e0cc8f8e837e115141dd4d4a916e0fc21e6
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20025
Reviewed-by: <anle.pan@nxp.com>