MA-20879-1 Revert "MA-19999 set state value in gadget HAL to notice HAL users"
This reverts commit 2a71d3feea9d3e8de48fb93ebc5978ead8b57ad3.

The reverted patch is related to a workaround of binding configfs with
dummy udc, that is a workaround in USB gadget HAL because adbd becomes a
mainline module.

Now adbd code is modified by Google to check a property
named"sys.usb.adb.disabled" to decide whether to pause the loop of
configuring the adbd ffs. So the workaround related patches are better
to be dropped.

Change-Id: Ibdacc151c93b39067248e604af8b6adee0b9fb60
Reviewed-on: http://androidsource.ap.freescale.net/project/20201
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>