MA-20891 Fix the issue that the displaymode of dcss mipi-to-hdmi change during hotplug
When processing hotplug event, the corresponding powermode will be set,
but SurfaceFlinger will only set power-on to primary display, so the adv
bridge is already power-on when it is plugged into hdmi, as a result, the
state of the ADV7511_REG_DDC_STATUS register is incorrect at this time, so
get edid will fail.

This patch won't notify the driver when the display is not connected and
the powermode is power-on.

Change-Id: Ib354641e9b43fdd28ef275bb43577c264f1d953f
Signed-off-by: zhai.he <zhai.he@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20232
Reviewed-by: zhang bo <bo.zhang@nxp.com>