MA-20953 Fix thhree resolutions c270 usbcamera camcorder preview is black
When the resolution is one of 160x120,640x360,800x600,
the decoder will fail to decod.
from bsp, The reason why it can’t be decoded is that
there are pad bytes(0xff) before the maker.
While the extra data may be not unresonable,
VSI vendors have not complied with this.
So use software to decode C270 USB camera.

Change-Id: I7629b381bdbcf2efe8e439c4d328da7a4bd0f0c2
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20612
Reviewed-by: fang hui <hui.fang@nxp.com>