MA-21231 refine fxos8700 scale based on the 6.1 fxos8700 driver change
fxos8700 driver merge several patches about data channel, scale and unit.

the current in_accel_scale defined in kernel is 0.000976, which is not proper,
the proper one should be 0.00976, use 0.00976 * in_accel_x(y/x)_raw, can
get reasonable values, when the device is stably put in the table, z value is
close to 9.8 while x, y value is close to 0, use scale 0.00976, the current acce
data is similar to the ones with the previous input framework.

mag scale is aligned with the one defined in kernel, in_magn_scale value is 0.001.

Change-Id: I9c9ff291cdadd77eff99dbda3ea986cbd9fbad18
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/21597
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>