MA-15746 Fix external camera preview freeze issue
On 8mq, c930 usb camera was decoded by hardware decoder,
when preview or picture, will meet decoder timeout issue,
but actually the decoding is done.
The issue is that the signal sent after the decoder DQBUF
is not received by the blocked wait thread, so the buff
is not returned normally, resulting in the buff nums decrease
and eventually exhausted, caused the preview freeze.
Return the output buff when timeout to ensure normal QBUF
next time.

Change-Id: I39ccc8086870004263eae604f864a180ec3e4e7c
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/21610
Reviewed-by: Hui Fang <hui.fang@nxp.com>