MA-21354-2 add the AIDL version of usb gadget HAL
Modify the file names to change "example" to "imx".

Change the user and group info in .rc file to be the same as the .rc
file for HIDL version HAL, or the HAL cannot access configfs files with
permission issue.

The license is removed because the license module defined under
hardware/interfaces cannot be accessed in this repo.

Modified the code with reference to the original HIDL implementation.

TEST: evk_8mn, switch gadget function between adb, mtp+adb,
    tethering+adb

    also modified the repo device/nsp/ to integrate this new HAL and set
    selinux label for the HAL binary.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I2b2dd0fff422fca3f9d2a3b39e672b353c48a632
Reviewed-on: http://androidsource.nxp.com/project/22329
Reviewed-by: Elven Wang <elven.wang@nxp.com>