MA-21354-4 modify the default AIDL usb HAL to adapt to imx
Modify the file names to change "example" to "imx".

Change the user and group info in .rc file to be the same as the .rc
file for HIDL version HAL, or the HAL cannot access configfs files with
permission issue.

The license is removed because the license module defined under
hardware/interfaces cannot be accessed in this repo.

Modified the code with reference to the original HIDL implementation
and the implementation of gs101 device.

Change-Id: Ic282b676ee8db982cead8bf0db1a970438870922
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/22331
Reviewed-by: Elven Wang <elven.wang@nxp.com>