MA-21420 Fix the green flash with dual display on 8mq
Generally, the output preview buffers are cycle used,
and to adapt for Camera2 apk, we skipped the resize action
when taking picture, in this case, the buffer content
was always the previously valid data.
But when with dual display, a new output preview buffer may
be used when taking picture, so green screen appears.
Not skip the process if there is a new dst phy buffer used.

Change-Id: I99dedfd90b1a374eaca835803293c23208655f59
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/22054
Reviewed-by: Hui Fang <hui.fang@nxp.com>