MA-20694 Fix the C270 UsbCamera preview abnormal issue
Some specific C270 is preview abnormal on 8mq,
8mq is hardware decoding for usb cameras, named vsi_v4l2dec.
C270 whose cardName is 'UVC Camera' is not ok,
C270 whose cardName is 'C270 HD WEBCAM' is ok, which may be
different production batches of the manufacturer.
Since from andy's email early this year: the old Logitech
C270(720P) are not supported, now we use software decoding
for C270 to keep the camera in use.

Change-Id: I2edd355abde1a3eb4a6cace1d747697ba834b877
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/22361
Reviewed-by: Anle Pan <anle.pan@nxp.com>