MA-20953 Fix thhree resolutions c270 usbcamera camcorder preview is black
When the resolution is one of 160x120,640x360,800x600,
the decoder will fail to decod.
from bsp, The reason why it can¡¯t be decoded is that
there are pad bytes(0xff) before the maker.
While the extra data may be not unresonable,
VSI vendors have not complied with this.
So use software to decode C270 USB camera.

Change-Id: Ie95ee3f646de0f801f1884c6d4a6a7528ceeb54e
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/22363
Reviewed-by: Anle Pan <anle.pan@nxp.com>