MA-15746 Fix external camera preview freeze issue
On 8mq, c930 usb camera was decoded by hardware decoder,
when preview or picture, will meet decoder timeout issue,
but actually the decoding is done.
The issue is that the signal sent after the decoder DQBUF
is not received by the blocked wait thread, so the buff
is not returned normally, resulting in the buff nums decrease
and eventually exhausted, caused the preview freeze.
Return the output buff when timeout to ensure normal QBUF
next time.

Change-Id: Ia838754fe1bc2aca916fc5cea4e973cb47f0be3c
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/22364
Reviewed-by: Anle Pan <anle.pan@nxp.com>