MA-21500 C920 preview black on 8qxp
Sometimes one v4l2 mjpg stream can't be decoded to one yuv frame.
so the output thread call processCaptureRequestError() and go to
next loop.

But it forget to call waitForBufferRequestDone(). It will lead
issue of "BufferRequestThread does not support more than one concurrent request!"
in next loop. Also the requested buffers in current loop are not return
back to framework.

Change-Id: Ic3218b637514280d58e52e2547b535a6434a24fa
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/22377
Reviewed-by: Anle Pan <anle.pan@nxp.com>