MA-21546 fix Gralloc5 validateBufferSize checking failure in CTS test
These failed cases are found when include stable-C version mapper which is used
in Gralloc5. Gralloc5 check the bufferSize of allocated buffer. It get the
requested pixel format through mapper and compare it with the one allocated.
The format of Memory structure is different from requested one. But the
fslFormat is the same one as requested pixel format.

Change-Id: If3b9d1507ae02265879fd2fa135dd12185d1c28c
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23206
Reviewed-by: Jessie Hao <juan.hao@nxp.com>