MA-21797-20 fix display blank issue when USB mouse cursor disappear
Found that sometimes the display will be blank and unblank again when UI update.
It usually happen when USB mouse cursor disappear.
The root cause is that the framework destroy some layer but the presentOrValidate
API is not completed and try to get the destroied layer. It will print the error
log like below:
11-14 02:53:24.760   456   533 E NXPHWC  : getLayer Unknown layer:35
11-14 02:53:24.770   485   485 E HWComposer: getDeviceCompositionChanges: presentOrValidate failed for display 4619827259835644672: BadLayer (3)
11-14 02:53:24.770   485   485 E CompositionEngine: chooseCompositionStrategy failed for EMU_display_0: -2147483648 (Unknown error -2147483648)

Because the presentOrValidate API cost too much time and cause such conflict.
Need to move some operation to present API to avoid such issue.

Change-Id: I3b07bad2d4af49865e428f9ac15a00228c53fe3d
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23354
Reviewed-by: Jessie Hao <juan.hao@nxp.com>