MA-21911 Fix C930 screen shakes and tears in resolution 424x240 issue
The issue happened with resolution 424x240.
The reason is that the outLayout data required by
halBuf.bufPtr has requirements for stride,
which was designed by gralloc according to the format,
and for nv12, it was 16-bit alignment.
Camera Hal did not take the dstStride into account when CSC.
Now handle it correctly to fix the issue.

Change-Id: I1617a07f32be137e3f58cd4d85cbf9be7623965d
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23422
Reviewed-by: Hui Fang <hui.fang@nxp.com>