MA-21921 Fix 8mp isp camera preview green issue when stress test
Issue happened after v4l2 DQBUF blocked, in this case,
will restart the 'isp_media_server', and will
reconfigured some parameters to recover the function,
when the service startup was not fully completed,
the subsequent configuration may cause extra errors.

Wait 1s before the reconfiguration to ensure that
the service is fully started.

Change-Id: I6004558a4a5ebf94b5e3b47016425adf2b51033a
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23465
Reviewed-by: Hui Fang <hui.fang@nxp.com>