MA-21797-43 Add poll thread to bring up display when DRM is ready
New PollThread is created when no client available and DummyClient used.
This thread will use epoll to check /dev/dri folder for any change. Then it
will create DrmClient and create DrmDisplay accordingly. The DummyClient will
be erased if the baseId of new DrmClient is the same as baseId of DummyClient
(it happens for most time). Only if the new DrmClient is mxsfb-drm(imx8mq-lcdif),
The DummyClient is kept.

Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Change-Id: I9dd36b11047f772c579c58d200014337fb58cfa4
Reviewed-on: http://androidsource.nxp.com/project/23754
Reviewed-by: Jessie Hao <juan.hao@nxp.com>