MA-22037 hwcomposer3: adjust amount of cached buffer to avoid secure heap not enough issue
The secure heap size is limited on most of boards. The framebuffer is allocated
from secure heap when playback secure video and use 2d composition. When the
amount of cached buffer is too large, the secure heap may be use up because of
old framebuffer is not freed in the secure and nonsecure video switching case.
Here limit the amount of cached buffers to 3 for framebuffer, when new buffer
added in, the old one will be dropped. When Overlay plane used, each overlay
plane can cache 8 buffers at most.

Change-Id: I4ceeab259bc6ac15ab9ac10fbf4417caf60d2e9f
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23808
Reviewed-by: Zhai He <zhai.he@nxp.com>
Reviewed-by: Jessie Hao <juan.hao@nxp.com>